requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.18.0  # For testing async functions